Interpolated line theme element

Interpolates a line.

element_line_seq(
  colour = NULL,
  size = NULL,
  linetype = NULL,
  lineend = NULL,
  color = NULL,
  arrow = NULL,
  inherit.blank = FALSE,
  fun = defer,
  n = 50
)

Arguments

colour
Line/border colour. Color is an alias for colour.

size
Line/border size in mm; text size in pts.

linetype
Line type. An integer (0:8), a name (blank, solid, dashed, dotted, dotdash, longdash, twodash), or a string with an even number (up to eight) of hexadecimal digits which give the lengths in consecutive positions in the string.

lineend
Line end Line end style (round, butt, square)

color
Line/border colour. Color is an alias for colour.

arrow
Arrow specification, as created by grid::arrow()

inherit.blank
Should this element inherit the existence of an element_blank among its parents? If TRUE the existence of a blank element among its parents will cause this element to be blank as well. If FALSE any blank parent element will be ignored when calculating final element state.

fun
A function to apply to the interpolated lines

n
An integer of length one indicating how many points to interpolate.

x
A single number specifying size relative to parent element.

Value

A list of the class element_line_seq

Examples

element_line_seq()

#> List of 8
#>  $ colour       : NULL
#>  $ size         : NULL
#>  $ linetype     : NULL
#>  $ lineend      : NULL
#>  $ arrow        : logi FALSE
#>  $ inherit.blank: logi FALSE
#>  $ n            : num 50
#>  $ fun          :function (...)  
#>  - attr(*, "class")= chr [1:3] "element_line_seq" "element_line" "element"