Multicolour lines

Interpolates lines and colour vectors, creating a gradient-like line. The element_*_multicolour() functions are convenience wrappers around element_*_seq().

multicolour()

element_line_multicolour(colour = rainbow(10), ...)

element_rect_multicolour(colour = rainbow(10), ...)

Arguments

colour
A character vector with colour names.

...
Arguments passed on to element_line_seq, element_rect_seq

fun

A function to apply to the interpolated lines

n

An integer of length one indicating how many points to interpolate.

size

Line/border size in mm; text size in pts.

linetype

Line type. An integer (0:8), a name (blank, solid, dashed, dotted, dotdash, longdash, twodash), or a string with an even number (up to eight) of hexadecimal digits which give the lengths in consecutive positions in the string.

lineend

Line end Line end style (round, butt, square)

color

Line/border colour. Color is an alias for colour.

arrow

Arrow specification, as created by grid::arrow()

inherit.blank

Should this element inherit the existence of an element_blank among its parents? If TRUE the existence of a blank element among its parents will cause this element to be blank as well. If FALSE any blank parent element will be ignored when calculating final element state.

sides

A character of length one containing any of "t", "l", "b" and/or "r". If these letters are present, edges are drawn at the top (t), left (l), bottom (b) or right (r), respectively. Including all these letters will draw all rectangle edges (default), whereas including none of these letters will not draw edges.

Value

A function.

Details

The multicolour function is a function factory that produces a function, that subsequently can be used to gradient-colour points.

For the element_rect_multicolour() the gradient is applied to each side seperately.

Examples

element_line_multicolour()

#> List of 8
#>  $ colour       : chr [1:10] "#FF0000FF" "#FF9900FF" "#CCFF00FF" "#33FF00FF" ...
#>  $ size         : NULL
#>  $ linetype     : NULL
#>  $ lineend      : NULL
#>  $ arrow        : logi FALSE
#>  $ inherit.blank: logi FALSE
#>  $ n            : num 50
#>  $ fun          :function (x, y, colour, id, n)  
#>  - attr(*, "class")= chr [1:3] "element_line_seq" "element_line" "element"